+++ database/migrations/2025_04_26_151119_create_tasks_table.php
@@ -11,19 +11,21 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('tasks', function (Blueprint $table) {
-            $table->id();
-            $table->string('title'); // Task title
-            $table->text('description')->nullable(); // Task description
-            $table->timestamp('start_time')->nullable(); // Initiation time
-            $table->timestamp('end_time')->nullable(); // Finalization time
-            $table->enum('status', ['pending', 'in-progress', 'completed'])->default('pending'); // Task status
-            $table->foreignId('user_id')->constrained()->onDelete('cascade'); // User who created the task
-            $table->foreignId('contract_id')->nullable()->constrained('contracts')->onDelete('cascade');
-            $table->boolean('is_billable')->default(false);
-            $table->decimal('billable_hours', 8, 2)->nullable(); // Hours worked on this task
-            $table->timestamps(); // Created at and updated at
-        });
+        if (!Schema::hasTable('tasks')) {
+            Schema::create('tasks', function (Blueprint $table) {
+                $table->id();
+                $table->string('title'); // Task title
+                $table->text('description')->nullable(); // Task description
+                $table->timestamp('start_time')->nullable(); // Initiation time
+                $table->timestamp('end_time')->nullable(); // Finalization time
+                $table->enum('status', ['pending', 'in-progress', 'completed'])->default('pending'); // Task status
+                $table->foreignId('user_id')->constrained()->onDelete('cascade'); // User who created the task
+                $table->foreignId('contract_id')->nullable()->constrained('contracts')->onDelete('cascade');
+                $table->boolean('is_billable')->default(false);
+                $table->decimal('billable_hours', 8, 2)->nullable(); // Hours worked on this task
+                $table->timestamps(); // Created at and updated at
+            });
+        }
     }
 
     /**
